FEATURE: consolidates maps data/types into one file
