+++ libs/spirit-islander/util/src/lib/data/maps.ts
@@ -1,4 +1,11 @@
-import type { Map } from '../types/game/maps';
+import type { DifficultyOption } from '../types/game/difficulty';
+import type { ExpansionOption } from '../types/game/expansions';
+
+export interface Map
+  extends DifficultyOption<MapName>,
+    ExpansionOption<MapName> {}
+
+export type MapName = 'Balanced' | 'Thematic';
 
 export const MAPS: readonly Map[] = [
   {
