+++ libs/spirit-islander/util/src/lib/types/combo.interface.ts
@@ -1,4 +1,4 @@
-import type { Map } from './game/maps';
+import type { Map } from '../data/maps';
 import type { AdversaryLevel } from './game/adversaries';
 import type { Scenario } from '../data/scenarios';
