+++ libs/spirit-islander/util/src/lib/types/config.interface.ts
@@ -1,6 +1,6 @@
 import type { Players } from './game/players';
 import type { Difficulty } from './game/difficulty';
-import type { MapName } from './game/maps';
+import type { MapName } from '../data/maps';
 import type { BalancedBoardName } from '../data/boards';
 import type { ExpansionName } from './game/expansions';
 import type { SpiritName } from '../data/spirits';
