+++ libs/spirit-islander/util/src/lib/types/game-setup.interface.ts
@@ -4,7 +4,7 @@ import type { Spirit } from '../data/spirits';
 import type { ExpansionName } from './game/expansions';
 import type { AdversaryLevel } from './game/adversaries';
 import type { Scenario } from '../data/scenarios';
-import type { Map } from './game/maps';
+import type { Map } from '../data/maps';
 import type { Board } from '../data/boards';
 
 /**
