--- libs/spirit-islander/util/src/lib/types/game/maps.ts
@@ -1,8 +1,0 @@
-import type { DifficultyOption } from './difficulty';
-import type { ExpansionOption } from './expansions';
-
-export type MapName = 'Balanced' | 'Thematic';
-
-export interface Map
-  extends DifficultyOption<MapName>,
-    ExpansionOption<MapName> {}
